% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asp_elev.R
\name{asp_elev_plot}
\alias{asp_elev_plot}
\title{Polar Scatter Plot of NASIS Site Data: Aspect, Elevation, and Ecologial Site}
\usage{
asp_elev_plot()
}
\value{
a polar scatterplot of site aspect and elevation - up to 10 ecosites
}
\description{
Polar Scatter Plot of NASIS Site Aspect and Elevation. Utilizes elevation, aspect, and site ecological site history data from NASIS site tables. Desired sites to plot need to be loaded into your NASIS selected set.
}
\examples{
asp_elev_plot()
}
