% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_veg_df.R
\name{create_veg_df}
\alias{create_veg_df}
\title{Create formatted vegetation dataframe}
\usage{
create_veg_df(
  from = c("web_report", "SS", "static"),
  ecositeid = NULL,
  static_location = NULL,
  best_vegplot = TRUE
)
}
\arguments{
\item{from}{origin of vegetation data. "web_report" provides public data in NASIS. "SS" uses the users NASIS selected set (NASIS connection required).
"static" uses a local .sqlite NASIS database.}

\item{ecositeid}{ecositeid of interest - required when using from = "web_report". If used with from = "SS" or from = "static", limits returns to only ecositeid provided.}

\item{static_location}{file path to static NASIS .sqlite database - typically generated using \code{ecositer::vegStaticNASIS()}}
}
\value{
vegetation dataframe fro ecological site data
}
\description{
This function provides access to vegetation data from multiple sources and returns the data in the same format. No QC is performed in this function.
}
\examples{

my_formatted_veg_df <- formatted_veg_df(static_location = "C:/Users/Nathan.Roe/Documents/SEKI/CA792_veg_data.sqlite")
head(formatted_veg_df)

}
