% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosite_similarity_list.R
\name{ecosite_similarity_pairwise}
\alias{ecosite_similarity_pairwise}
\title{Ecosite similarity list}
\usage{
ecosite_similarity_pairwise(veg_df, dist_pair)
}
\arguments{
\item{veg_df}{a properly formatted vegetation dataframe (\link[ecositer]{formatted_veg_df}).
The data should be QCed to ensure that it meets minimum data requirements (\link[ecositer]{QC_vegplots})}

\item{dist_pair}{a distance}
}
\value{
a list of ecosite similarities
}
\description{
Ecosite similarity list
}
\examples{
ecositer::ecosite_similarity_pairwise(veg_df = my_veg_df, dist_pair = my_dist)
}
