% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecositer-package.R
\docType{package}
\name{ecositer-package}
\alias{ecositer}
\alias{ecositer-package}
\title{ecositer: ecositer: tools for the analysis, standardization, and automation of NRCS ecological data}
\description{
A collection of functions for Natural Resource Conservation Service (NRCS) ecologists. Many of these functions build off of the AQP (https://ncss-tech.github.io/AQP/) R packages to provide improved functionality for ecologists.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/natearoe/ecositer}
  \item Report bugs at \url{https://github.com/natearoe/ecositer/issues}
}

}
\author{
\strong{Maintainer}: 

Other contributors:
\itemize{
  \item Nathan Roe \email{nateroe@gmail.com}
}

}
\keyword{internal}
