% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatted_veg_df.R
\name{formatted_veg_df}
\alias{formatted_veg_df}
\title{Create formatted vegetation dataframe}
\usage{
formatted_veg_df(
  SS = TRUE,
  static_location = NULL,
  cover_column = c("akstratumcoverclasspct", "speciescancovpct", "speciescomppct",
    "understorygrcovpct")
)
}
\arguments{
\item{SS}{TRUE or FALSE, should data be accessed from NASIS Selected Set?}

\item{static_location}{the location where a static NASIS database has been
saved.}

\item{cover_column}{NASIS column where vegetation cover data is stored}
}
\value{
a formatted vegetation dataframe suitable for analysis
}
\description{
Create formatted vegetation dataframe
}
\examples{

my_formatted_veg_df <- formatted_veg_df(static_location = "C:/Users/Nathan.Roe/Documents/SEKI/CA792_veg_data.sqlite")
head(formatted_veg_df)

}
