% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmds_ecosite.R
\name{nmds_ecosite}
\alias{nmds_ecosite}
\title{Run NMDS on vegetation data}
\usage{
nmds_ecosite(veg_summary, ecosite, pres_abs, nmds_dim, reduce_species)
}
\arguments{
\item{veg_summary}{veg_summary object where the \code{ecositer::veg_summary()} function was saved.}

\item{ecosite}{ecosite is the ecosite(s) you want to include in an ordination. If you are
interested in ordinating multiple ecosites together, concatenate the ecosites of interest
(e.g., \code{c("a", "b")})}

\item{pres_abs}{use TRUE to convert data to presence/absence. This is the more
robust option. If you have missing abundance values, presence/absence should be used.
If you have abundance data, use FALSE.}

\item{nmds_dim}{specify the number of dimensions you want in your NMDS. This value
will affect stress considerably.}

\item{reduce_species}{\code{NA} if you do not want to remove species. A numeric value,
interpretted as percent, specifying the minimum percentage of plots a species
must be present in to remain in the dataset.}
}
\value{
an NMDS object generated using the vegan package
}
\description{
\code{nmds_ecosite} runs an NMDS using the vegan package. NMDS has numerous settings
that can be adjusted, not all of which can be manipulated through the arguments
of this function. This function is intended to be used as a quick way of producing
ordinations. If you are interested in optimizing your NMDS, please do your own
research. At a minimum, you should understand stress and evaluate the stress
report in the results.
}
\examples{
formatted_veg <- formatted_veg_df(static_location = "C:/Users/Nathan.Roe/Documents/SEKI/CA792_veg.sqlite")
my_veg_summary <- ecositer::veg_summary(veg_df = formatted_veg)
my_nmds <- ecositer::nmds_ecosite(veg_summary = my_veg_summary, ecosite = "F022AK100CA", pres_abs = TRUE, nmds_dim = 2, reduce_species = NA)
}
