% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmds_taxa.R
\name{nmds_taxa}
\alias{nmds_taxa}
\title{Run NMDS for taxa of interest}
\usage{
nmds_taxa(veg_df, taxa, spp.scr.thrshld = 0.01)
}
\arguments{
\item{veg_df}{properly formatted vegetation dataframe}

\item{taxa}{taxa of interested (e.g., "Arenaria")}

\item{spp.scr.thrshld}{p-value threshold for species scores being displayed}
}
\value{
an ordination of plots in species space of all plots that contain the taxa of interest
}
\description{
This function is intended to provide a quick visualization of sites containing
a taxa of interest. It is intended to be used in assessing whether taxa can
be combined. For example, an ordination showing strong overlap in vegetative
communities between sites with "Arenaria" and "Arenaria kingii" might suggest
that these two taxa can be combined.
}
