% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pres_abs_conversion.R
\name{pres_abs_conversion}
\alias{pres_abs_conversion}
\title{Convert normal abundance data to presence/absence}
\usage{
pres_abs_conversion(veg_df)
}
\arguments{
\item{veg_df}{vegetation dataframe that will be converted to presence/absence.}
}
\value{
a presence/absence dataframe
}
\description{
\code{pres_abs_conversion} is an internal function that is used within the
\link[ecositer]{nmds_ecosite} function. Using presence/absence data will
allow you to use data from plots that do not have abundance data. In certain
datasets, this is an advantageous feature. When I have used NMDS on moderate
sized datasets (i.e., 100+ plots), I have found it to yield similar results
as abundance data. This could be different for small datasets.
}
