% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_best_vegplot_for_site.R
\name{QC_best_vegplot_for_site}
\alias{QC_best_vegplot_for_site}
\title{Keep best vegplot when multiple exist for site}
\usage{
QC_best_vegplot_for_site(veg_df)
}
\arguments{
\item{veg_df}{}
}
\value{
a vegetation dataframe with one vegplot per site
}
\description{
This function determines what sites have multiple vegplots and
keeps only the best vegplot. The best vegplot is decided based on
the number of records for the vegplot (i.e., the number of rows in the
vegplot table). This function is intended to
provide basic functionality to avoid gross misrepresentations (i.e.,
pseudo-replication of vegplot/site relationships). This function should
not be considered a replacement for careful review of vegplots. If,
there are multiple vegplots with the same number of records, this function
will error. This most commonly occurs when there are two vegplots with
identical data. The best solution in that case is to remove duplicated
vegplots from NASIS.
}
