% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_veg_completeness.R
\name{QC_completeness_criteria}
\alias{QC_completeness_criteria}
\title{Remove vegetation plots based on completeness criteria}
\usage{
QC_completeness_criteria(
  veg_df = NULL,
  min_total_records = NULL,
  min_unique_species = NULL,
  min_perc_to_species = NULL,
  min_perc_with_abund = NULL
)
}
\arguments{
\item{veg_df}{vegetation dataframe}

\item{min_species}{(\%) - sites must have >= min_species to be included}

\item{min_abund}{integer - sites must have >= min_abund species with abundance data to be included (e.g.,
entering 90 would include all sites with 90\% or more species with abundance data).
WARNING: If you keep sites that are missing abundance data and intend to run vegetation
analyses using abundance data, abundance values will have to be populated, otherwise that species record will be removed. If you run
analyses on presence/absence data, abundance data is not needed.}
}
\value{
vegetation dataframe meeting completeness criteria
}
\description{
Remove vegetation plots based on completeness criteria
}
\examples{
QC_completeness_criteria(veg_df = my_veg_df, min_unique_species = 5, min_perc_to_species = 70, min_perc_with_abund = 70)
}
\seealso{
\code{\link[=QC_veg_completeness]{QC_veg_completeness()}}
}
