% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_abundance.R
\name{QC_find_multiple_abundance}
\alias{QC_find_multiple_abundance}
\title{Find vegetation records with multiple abundances}
\usage{
QC_find_multiple_abundance(veg_df)
}
\arguments{
\item{veg_df}{vegetation dataframe}
}
\value{
dataframe of records with multiple abundance columns populated
}
\description{
This function returns records that use multiple abundance columns. This allows users
to determine the cause and determine a better solution than averaging.
}
\examples{
QC_find_multiple_abundance(veg_df = B100_veg)

}
\seealso{
\code{\link[=QC_aggregate_abundance]{QC_aggregate_abundance()}}
}
