% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_taxonomy.R
\name{QC_taxonomy}
\alias{QC_taxonomy}
\title{QC taxonomy}
\usage{
QC_taxonomy(veg_df)
}
\arguments{
\item{veg_df}{a properly formatted vegetation dataframe.}
}
\value{
vegetation dataframe with updated taxonomy according to USDA PLANTS.
}
\description{
QC taxonomy
}
\examples{
QC_taxonomy(veg_df = my_veg_df)
}
