% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_plants.R, R/QC_unifying_taxa.R
\name{QC_unify_taxa}
\alias{QC_unify_taxa}
\title{QC plants}
\usage{
QC_unify_taxa(veg_df)

QC_unify_taxa(veg_df)
}
\arguments{
\item{veg_df}{}
}
\value{
a dataframe of taxa recommended for unifying

a dataframe of taxa recommended for unifying
}
\description{
\code{QC_unify_data} returns a dataframe of taxa that could benefit from unifying. Any identification beyond species
(i.e., subspecies, variety, etc.) will be returned, as will identifications to genus level. Statistical analyses do
not recognize any similarity between Calamagrostis stricta and Calamagrostis stricta ssp. stricta, so if it is reasonable
to unify these to the same class (e.g., change all to Calamagrostis stricta), that is recommendable.

\code{QC_unify_data} returns a dataframe of taxa that should be QCed for opportunities to unify. There are two types of situations
that are evaluatied. The first is whether there are IDs beyond species (i.e., var or ssp) as well as to species.
For example, Pinus contorta and Pinus contorta var. murrayana. Users may want to unify these observations
(e.g., change all to Pinus contorta). The second situation is observations to genus as well as genus species. For example
IDs to Pinus and to Pinus contorta. In some instances, it will be possible to unify these records (e.g., at the end of a
project you determine there is only one species of a genus present). Statistical analyses do not recognize any relationship
between different classes, therefore these taxonomically similar situations are viewed by distance measures as completely
unrelated. Therefore, when reasonable, it is best to unify taxonomic observations.
}
\examples{
QC_plants(veg_df = ecositer::vegetation_dataframe) |> head()
QC_plants(veg_df = ecositer::vegetation_dataframe) |> head()
}
