% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_update_taxonomy.R
\name{QC_update_taxonomy}
\alias{QC_update_taxonomy}
\title{Update taxonomy according to USDA PLANTS}
\usage{
QC_update_taxonomy(veg_df)
}
\arguments{
\item{veg_df}{}
}
\value{
vegetation dataframe with updated taxonomy
}
\description{
Update taxonomy according to USDA PLANTS
}
\examples{
ecositer::QC_update_taxonomy(veg_df = CA792_veg)
}
