% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_veg_completeness.R
\name{QC_veg_completeness}
\alias{QC_veg_completeness}
\title{Summarize vegetation plot completeness}
\usage{
QC_veg_completeness(veg_df)
}
\arguments{
\item{veg_df}{}
}
\value{
a summary of vegetation plot completeness
}
\description{
This function calculates summary statistics for vegetation plots grouped by \code{vegplotiid}, specifically:
total_records - the total number of records associated with each \code{vegplotiid}
unique_species - The number of unique species recorded for each \code{vegplotiid}.
percent_to_species - the percentage of observations identified to the species level (determined by the presence of a space in \code{plantsciname}).
percent_with_abund - the percentage of observations with a non-NA value in the \code{pct_cover} column.

\code{QC_vegplots_remove_plots()} allows you to remove vegplots from your dataset
based on siteiids and parameters related to vegplot completeness.
}
\examples{
my_QC_vegplots <- QC_veg_completeness(veg_df = my_veg_df)
}
\seealso{
\code{\link[=QC_completeness_criteria]{QC_completeness_criteria()}}
}
