% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanned_data_organizer.R
\name{scanned_data_organizer}
\alias{scanned_data_organizer}
\title{Organize scanned datasheets into folders}
\usage{
scanned_data_organizer(
  scanned_data_path,
  directory,
  year_state_fips,
  site_number
)
}
\arguments{
\item{scanned_data_path}{the location of your scanned datasheets}

\item{directory}{the directory where you want the results stored}

\item{year_state_fips}{the year, state, and fips code for your site id
(e.g., "2023AK185")}

\item{site_number}{the site number for example in site id 2023AK185401, the
site_number is 401. site_number must be manually entered as a vector unless
you can extract the site_number from GPS point names or some other record
of the sites your sampled. If you cannot extract the site_number, you will
enter it following this format:
my_site_number <- c(401, 402, 405, 409, 410, 411)}
}
\value{
the result is a directory with folders for each site, within which
is a pdf of your datasheet named with the site name. A folder containing
all sites will also be created.
}
\description{
\code{scanned_data_organizer} is a tool to take scans of field data and organize it
into folders. The process starts with scanning your datasheets. All of
your datasheets can be scanned at once. The result of which will be a
single pdf file containing all of your datasheets. This function will
separate that pdf file into the individual sites that you sampled and
name those new pdfs using the site id. A folder will be created for each
site id, within which the datasheet will be stored. In addition, an
all_sites folder will be created where all of the pdfs will be stored
together.
}
\examples{
my_site_number <- c(100, 101, 102, 103, 123, 124, 125)
scanned_data_organizer(scanned_data_path = "S:/NRCS/XEROX SCANS/DOC057.pdf",
directory = "C:/Users/Nathan.Roe/Documents/Alaska2023/Willow_project",
year_state_fips = "2023AK185",
site_number = my_site_number)

}
