% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_level_soil_properties.R
\name{site_level_soil_properties}
\alias{site_level_soil_properties}
\title{Create site level properties from soil horizon data}
\usage{
site_level_soil_properties(
  soil_data,
  checkHzDepthLogic = TRUE,
  stopOnHzLogicFail = TRUE,
  byDepth = list(c(0, 25), c(0, 50))
)
}
\arguments{
\item{soil_data}{a soilDB SoilProfileCollection (SPC)}

\item{checkHzDepthLogic}{evaluate logic of horizon data. A warning or error will be issued depending on the stopOnHzLogicFail argument. aqp::checkHzDepthLogic() should be run on your data is logic errors exist and corrections should be made to data.}

\item{stopOnHzLogicFail}{will prevent this function from running if logic errors exist in horizon data.}

\item{byDepth}{a list of depth ranges to calculate properties across. Use NULL if depth ranges are not of interest.}
}
\value{
an SPC with ecologically relevant properties appended to the site df
}
\description{
Calculate properties from soil horizon data that will be appended to the site table of a
SoilProfileCollection. The entire soil profile is used. Additionally,the byDepth argument allows
for calculations to be made on specified depth ranges.
}
