% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_similarity_pairwise.R
\name{site_similarity_pairwise}
\alias{site_similarity_pairwise}
\title{Distance matrix to pairwise distance dataframe}
\usage{
site_similarity_pairwise(dist_matrix)
}
\arguments{
\item{dist_matrix}{}
}
\value{
a data frame of pairwise distances between plots
}
\description{
Distance matrix to pairwise distance dataframe
}
\examples{
site_similarity_pairwise(dist_matrix = my_dist)
}
