% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_soils.R
\name{site_soils}
\alias{site_soils}
\title{Site level soil properties}
\usage{
site_soils(pedon_data)
}
\arguments{
\item{pedon_data}{an aqp SoilProfileCollection of pedon data}
}
\value{
a site-level dataframe with aggregated soil properties
}
\description{
Site level soil properties
}
\examples{
site_soils(pedon_data = my_peds)
}
