% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_list.R
\name{species_list}
\alias{species_list}
\title{Species list}
\usage{
species_list(SS = TRUE, static_location = NULL)
}
\arguments{
\item{SS}{logical - whether to use access data in NASIS selected set}

\item{dsn}{}
}
\value{
a species list from your NASIS vegplots with modern taxonomies, growth habit, native status, and higher level taxa from PLANTS database
}
\description{
Species list
}
\examples{
species_list(SS = TRUE)
}
