% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_pedon_soil_properties.R
\name{summarize_pedon_soil_properties}
\alias{summarize_pedon_soil_properties}
\title{Summarize pedon soil properties}
\usage{
summarize_pedon_soil_properties(
  SS = TRUE,
  static_location = NULL,
  r_object = NULL,
  byDepth = list(c(0, 25), c(0, 50))
)
}
\arguments{
\item{r_object}{R object where SoilProfileCollection is stored}

\item{byDepth}{}
}
\description{
This function provides summarized soil properties for pedons. Summarized properties
are calculated for the entire profile as well as for user specified depth ranges.
Properties include texture, pH, fragment volume,
}
