% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing_veg.R
\name{summarize_veg}
\alias{summarize_veg}
\title{Summarize vegetation data by ecosite, state, and community phase}
\usage{
summarize_veg(
  veg_df,
  agg_abund = TRUE,
  update_taxonomy = TRUE,
  best_vegplot = TRUE,
  use_id = TRUE
)
}
\arguments{
\item{use_id}{logical whether to use state and phase id (i.e., numeric) or name}
}
\value{
list of summarized vegetation data
}
\description{
Summarize vegetation data by ecosite, state, and community phase
}
