% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_veg_by_ecosite.R
\name{summarize_veg_by_ecosite}
\alias{summarize_veg_by_ecosite}
\title{Summarize vegetation data by ecosite, state, and community phase}
\usage{
summarize_veg_by_ecosite(
  veg_df,
  agg_abund = TRUE,
  update_taxonomy = TRUE,
  best_vegplot = TRUE
)
}
\arguments{
\item{veg_df}{vegetation dataframe}

\item{agg_abund}{logical - aggregate abundance columns using \code{ecositer::QC_aggregate_abundance()}}

\item{update_taxonomy}{logical - update taxonomy from USDA PLANTS using \code{ecositer::QC_update_taxonomy()},}

\item{best_vegplot}{logical - use best vegplot when multiple for site using \code{ecositer::QC_best_vegplot_for_site()},}
}
\value{
list of summarized vegetation data
}
\description{
Summarize vegetation data by ecosite, state, and community phase
}
