% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_summary.R
\name{veg_summary}
\alias{veg_summary}
\title{Produce vegetation summary of ecosites and states/phases}
\usage{
veg_summary(veg_df)
}
\arguments{
\item{veg_df}{a properly formatted vegetation dataframe}
}
\value{
a nested list of vegetation summaries by ecosite and state/phase
}
\description{
Produce vegetation summary of ecosites and states/phases
}
\examples{
#' my_veg_summary <- veg_summary(veg_df = ecositer::vegetation_dataframe)
head(my_veg_summary)

}
