% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_summary_taxa.R
\name{veg_summary_taxa}
\alias{veg_summary_taxa}
\title{Vegetation summary for particular taxa}
\usage{
veg_summary_taxa(veg_df, taxa)
}
\arguments{
\item{veg_df}{properly formatted vegetation data frame}

\item{taxa}{taxa of interest, typically genus level (e.g., "Arabis"). This is used as a regular expression pattern.}
}
\value{
a vegetation summary for the taxa of interest
}
\description{
Vegetation summary for particular taxa
}
\examples{
ecositer::veg_summary_taxa(veg_df = my_veg_df, taxa = "Arabis")
}
