% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_example_data.R
\name{access_example_data}
\alias{access_example_data}
\title{Accessing example data}
\usage{
access_example_data()
}
\value{
the location of example data
}
\description{
\code{access_example_data()}is an internal function that helps function examples
and vignettes to access datasets
}
\examples{
access_example_data()


}
