% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatted_veg_df.R
\name{formatted_veg_df}
\alias{formatted_veg_df}
\title{Create formatted vegetation dataframe}
\usage{
formatted_veg_df(static_location)
}
\arguments{
\item{static_location}{the location where a static NASIS database has been
saved.}
}
\value{
a formatted vegetation dataframe suitable for analysis
}
\description{
Create formatted vegetation dataframe
}
\examples{

formatted_veg_df(static_location = "C:/Users/Nathan.Roe/Documents/SEKI/CA792_veg_data.sqlite")


}
