% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmds_plot.R
\name{nmds_plot}
\alias{nmds_plot}
\title{NMDS plot}
\usage{
nmds_plot(static_location, nmds, veg_summary)
}
\arguments{
\item{static_location}{path to static nasis SQlite.}

\item{nmds}{object created from \code{ecositer::nmds_ecosite}}

\item{veg_summary}{objected created from \code{ecositer::veg_summary}}
}
\value{
interactive NMDS plot
}
\description{
\code{nmds_plot} produces and interactive NMDS from the object created with \code{ecositer::nmds_ecosite}. The function allows the user to define several properties,
such as where to use pres/abs data, the number of dimensions, and whether to remove rare species.
}
