% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_plots_by_site.R
\name{number_plots_by_site}
\alias{number_plots_by_site}
\title{Number of plots by ecosite}
\usage{
number_plots_by_site(static_location)
}
\arguments{
\item{static_location}{the location where a static NASIS database has been
saved.}
}
\value{
a dataframe describing the number of pedons and vegplots recorded
for each ecosite
}
\description{
\code{number_plots_by_ecosite} describes the number of pedons and vegplots that
have been recorded for each ecosite. It also lists the states/phases
associated with each ecosite.
}
\examples{

my_plots_by_site <- number_plots_by_site(static_location = "C:/Users/Nathan.Roe/Documents/SEKI/vegplotdata2.sqlite")
head(my_plots_by_site)

}
