% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_abundance.R
\name{QC_aggregate_abundance}
\alias{QC_aggregate_abundance}
\title{Aggregate abundance columns}
\usage{
QC_aggregate_abundance(veg_df, fail_on_dup = FALSE)
}
\arguments{
\item{veg_df}{vegetation dataframe from \code{ecositer::create_veg_df()}}

\item{fail_on_dup}{logical - whether the function should fail if any records
have multiple abundance columns populated.}
}
\value{
dataframe with one abundance column
}
\description{
This function is used to aggregate abundance data into one column (pct_cover). Abundance can come from four columns:
"akstratumcoverclasspct", "speciescancovpct", "speciescomppct", and "understorygrcovpct". If the same record uses multiple abundance
columns, pct_cover is calculated as the average of those values. When only one abundance column is used for a record,
that value is used for pct_cover.
}
\examples{
QC_aggregate_abundance(veg_df = B100_veg)
}
\seealso{
\code{\link[=QC_find_multiple_abundance]{QC_find_multiple_abundance()}}
}
