% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_best_vegplot_for_site.R
\name{QC_best_vegplot_for_site}
\alias{QC_best_vegplot_for_site}
\title{Keep best vegplot when multiple exist for site}
\usage{
QC_best_vegplot_for_site(veg_df)
}
\arguments{
\item{veg_df}{}
}
\value{
a vegetation dataframe with one vegetation plot per site
}
\description{
This function determines what sites have multiple vegetation plots and
keeps only the best. The best vegetation plot is decided based on
the number of records for the vegetation plot (i.e., the number of rows in the
vegetation plot table). This function is intended to
provide basic functionality to avoid gross misrepresentations (i.e.,
pseudo-replication of vegetation plot/site relationships). This function should
not be considered a replacement for careful review of vegetation plots. If,
there are multiple vegetation plots with the same number of records, this function
will error. This most commonly occurs when there are two vegetation plots with
identical data. The best solution in that case is to remove duplicated
vegetation plots from NASIS.
}
