% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_location_data.R
\name{QC_location_data}
\alias{QC_location_data}
\title{QC site location data}
\usage{
QC_location_data(veg_df)
}
\arguments{
\item{veg_df}{vegetation dataframe, commonly created using \code{ecositer::create_veg_df()}}
}
\value{
dataframe uneeded coordinate format columns removed
}
\description{
QC site location data
}
\examples{
QC_location_data(veg_df = my_veg_df)
}
