% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_unifying_taxa.R
\name{QC_unify_taxa}
\alias{QC_unify_taxa}
\title{QC taxonomies for unifying}
\usage{
QC_unify_taxa(veg_df)
}
\arguments{
\item{veg_df}{}
}
\value{
a dataframe of taxa recommended for unifying
}
\description{
\code{QC_unify_data} returns a dataframe of taxa that should be QCed for opportunities to unify. There are two types of situations
that are evaluatied. The first is whether there are IDs beyond species (i.e., var or ssp) as well as to species.
For example, Pinus contorta and Pinus contorta var. murrayana. Users may want to unify these observations
(e.g., change all to Pinus contorta). The second situation is observations to genus as well as genus species. For example
IDs to Pinus and to Pinus contorta. In some instances, it will be possible to unify these records (e.g., at the end of a
project you determine there is only one species of a genus present). Statistical analyses do not recognize any relationship
between different classes, therefore these taxonomically similar situations are viewed by distance measures as completely
unrelated. Therefore, when reasonable, it is best to unify taxonomic observations.
}
\examples{
QC_plants(veg_df = ecositer::vegetation_dataframe) |> head()
}
