% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_unifying_taxa.R
\name{QC_unify_taxa}
\alias{QC_unify_taxa}
\title{Identify Taxa for Unification}
\usage{
QC_unify_taxa(veg_df)
}
\arguments{
\item{veg_df}{vegetation dataframe}
}
\value{
a dataframe of taxa recommended for unifying
}
\description{
This function returns a data frame of plant taxonomies that should be evaluated for unification. Unifying taxa means lumping
multiple distinct taxa together. Statistical analyses do not recognize any relationship
between different taxonomic classes, therefore taxonomically similar situations are viewed by distance measures as
unrelated, despite potentially having very close relationship. Therefore, it is more meaningful to unify taxa rather than preserving
distinctions between closely related taxa.

There are two types of situations that are evaluated:
\enumerate{
\item \strong{Infraspecific Species Variation}: Cases where observations exist both at the species and infraspecific levels (e.g., \emph{Pinus contorta} and \emph{Pinus contorta var. murrayana}). Users may wish to collapse these to the species level.
\item \strong{Mixed Genus and Species-Level Observations}: Records identified to genus (e.g., \emph{Pinus}) alongside specific species within that genus (e.g., \emph{Pinus contorta}). In data sets where there are genus level observations and
either a single species, or species with similar ecological function, users may collapse to the genus level.
}
}
\examples{
QC_unify_taxa(veg_df = SEKI_NP_veg)
}
