% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_vegplots.R
\name{QC_vegplots_summary}
\alias{QC_vegplots_summary}
\alias{QC_vegplots_remove_plots}
\title{QC vegplots}
\usage{
QC_vegplots_summary(veg_df)

QC_vegplots_remove_plots(
  veg_df = NULL,
  my_QC_vegplots_summary = NULL,
  min_species = NULL,
  min_perc_abund = NULL,
  site_ids = NULL
)
}
\arguments{
\item{veg_df}{veg_df created using \code{ecositer::formatted_veg_df()}}

\item{my_QC_vegplots_summary}{results of \code{ecositer::my_QC_vegplots_summary()}}

\item{min_species}{site must have >= N species to be included}

\item{min_abund}{sites must have >= N\% species with abundance data to be included (e.g.,
entering 90 would include all sites with 90\% or more species with abundance data).
WARNING: If you keep sites that are missing abundance data and intend to run vegetation
analyses using abundance data, abundance values will have to be populated. If you run
analyses on presence/absence data, abundance data is not needed.}

\item{site_id}{site_ids to be removed as defined by user}
}
\value{
\code{QC_vegplots_summary()} returns a summary of vegetation plot sampling detail

\code{QC_vegplots_remove_plots()} returns a veg_df satisfying the defined
requirements of completeness
}
\description{
\code{QC_vegplots_summary()} produces a summary regarding the completeness of
vegplots, including the species richness in each plot and the percent of
species with abundance data. This information can be used to remove vegplots
from analyses due to lack of completeness.

\code{QC_vegplots_remove_plots()} allows you to remove vegplots from your dataset
based on site_ids and parameters related to vegplot completeness.
}
\examples{
my_QC_vegplots <- QC_vegplots(veg_df = my_veg_df)
# example
}
