% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizon_to_site.R
\name{site_level_soil_properties}
\alias{site_level_soil_properties}
\title{Create site level properties from soil horizon data}
\usage{
site_level_soil_properties(soil_data)
}
\arguments{
\item{soil_data}{a soilDB SoilProfileCollection (SPC)}
}
\value{
an SPC with ecologically relevant properties appended to the site df
}
\description{
Create site level properties from soil horizon data
}
