% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_summary.R
\name{veg_summary}
\alias{veg_summary}
\title{Produce vegetation summary of ecosites and states/phases}
\usage{
veg_summary(veg_df)
}
\arguments{
\item{veg_df}{a properly formatted vegetation dataframe}
}
\value{
a nested list of vegetation summaries by ecosite and state/phase
}
\description{
\code{veg_summary()} takes a \emph{properly formatted vegetation dataframe} and produces summaries
of vegetation for ecosites and states/phases. Vegetation data originates from the
NASIS vegplot table. The BLANK function should be used to manipulate raw NASIS
vegplot data into a \emph{properly formatted vegetation dataframe}
}
\examples{

veg_summary(veg_df = ecositer::vegetation_dataframe)


}
